+++ src/RavenNest.DataModels/ClanSkill.cs
@@ -8,6 +8,6 @@ namespace RavenNest.DataModels
         private Guid clanId; public Guid ClanId { get => clanId; set => Set(ref clanId, value); }
         private Guid skillId; public Guid SkillId { get => skillId; set => Set(ref skillId, value); }
         private int level; public int Level { get => level; set => Set(ref level, value); }
-        private decimal experience; public decimal Experience { get => experience; set => Set(ref experience, value); }
+        private double experience; public double Experience { get => experience; set => Set(ref experience, value); }
     }
 }
